Split compact cmd
