+++ cmd/util-db/main.go
@@ -18,6 +18,7 @@ package main
 
 import (
 	"github.com/0xsoniclabs/aida/cmd/util-db/clone"
+	"github.com/0xsoniclabs/aida/cmd/util-db/compact"
 	"log"
 	"os"
 
@@ -35,7 +36,7 @@ var UtilDbApp = cli.App{
 	Copyright: "(c) 2022 Fantom Foundation",
 	Commands: []*cli.Command{
 		&clone.Command,
-		&db.CompactCommand,
+		&compact.Command,
 		&db.ExtractEthereumGenesisCommand,
 		&db.MergeCommand,
 		&db.UpdateCommand,
